Move controller to dependency injection
