+++ Admin/FormAdmin.php
@@ -22,9 +22,6 @@ use Sulu\Component\Security\Authorization\PermissionTypes;
 use Sulu\Component\Security\Authorization\SecurityCheckerInterface;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
 
-/**
- * Generated by https://github.com/alexander-schranz/sulu-backend-bundle.
- */
 class FormAdmin extends Admin
 {
     const LIST_VIEW = 'sulu_form.list';
