+++ Controller/FormController.php
@@ -28,9 +28,6 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
-/**
- * Generated by https://github.com/alexander-schranz/sulu-backend-bundle.
- */
 class FormController extends AbstractRestController implements ClassResourceInterface, SecuredControllerInterface
 {
     /**
