+++ Manager/FormManager.php
@@ -18,9 +18,6 @@ use Sulu\Bundle\FormBundle\Entity\FormTranslation;
 use Sulu\Bundle\FormBundle\Entity\FormTranslationReceiver;
 use Sulu\Bundle\FormBundle\Repository\FormRepository;
 
-/**
- * Generated by https://github.com/alexander-schranz/sulu-backend-bundle.
- */
 class FormManager
 {
     /**
