+++ Resources/config/services.xml
@@ -229,13 +229,25 @@
         </service>
 
         <!-- Form Website Controller -->
-        <service id="Sulu\Bundle\FormBundle\Controller\FormWebsiteController">
+        <service id="sulu_form.form_website_controller"
+            class="Sulu\Bundle\FormBundle\Controller\FormWebsiteController"
+            public="true">
             <tag name="container.service_subscriber" />
             <tag name="controller.service_arguments" />
             <tag name="sulu.context" context="website"/>
+
             <call method="setContainer">
                 <argument type="service" id="Psr\Container\ContainerInterface" />
             </call>
+        </service>
+
+        <!-- Form Website Controller -->
+        <service id="sulu_form.form_token_controller"
+                 class="Sulu\Bundle\FormBundle\Controller\FormTokenController"
+                 public="true">
+            <argument type="service" id="security.csrf.token_manager"/>
+
+            <tag name="sulu.context" context="website"/>
         </service>
 
         <!-- Dynamic List Metadata Loader -->
