+++ Resources/views/themes/basic.html.twig
@@ -21,7 +21,7 @@
 
 {# CSRF Token over ESI #}
 {%- block csrf_token_widget %}
-    {{ render_esi(controller('Sulu\\Bundle\\FormBundle\\Controller\\FormWebsiteController::tokenAction', {
+    {{ render_esi(controller('Sulu\\Bundle\\FormBundle\\Controller\\FormTokenController::tokenAction', {
         'form': form.parent.vars.name,
         'html': true,
         _requestAnalyzer: false }))
